requirements.txt
agno>=1.4.0
openai>=1.30.0
python-dotenv>=1.0.0
rich>=13.7.0
requests>=2.31.0
